reorganize files for blockchain example and kyc components
