+++ example-apps/src/features/gated-nfts/GatedNFTPage.tsx
@@ -1,8 +1,8 @@
 import { Inter } from "next/font/google";
 import styles from "./client.module.css";
-import { KYCClient } from "./identity/KYCClient";
+import { KYCClient } from "../kyc-widget/KYCClientEIP155";
 import { useState } from "react";
-import { GatedNFT } from "./blockchain/GatedNFT";
+import { GatedNFT } from "./components/GatedNFT";
 
 const inter = Inter({ subsets: ["latin"] });
