+++ example-apps/src/features/gated-nfts/components/DisplayMintedNFTs.tsx
