+++ example-apps/src/features/gated-nfts/utils/blockchain.schema.ts
