+++ example-apps/src/features/gated-nfts/utils/getContractAddress.ts
