+++ example-apps/src/pages/gated-nft.tsx
@@ -1,4 +1,4 @@
 import dynamic from "next/dynamic";
-import { GatedNFTPage } from "@/features/gated-nfts/gated-nfts/GatedNFTPage";
+import { GatedNFTPage } from "@/features/gated-nfts/GatedNFTPage";
 
 export default dynamic(() => Promise.resolve(GatedNFTPage), { ssr: false });
